+++ adapter/src/main/java/ee/cyber/cdoc2/server/adapter/exception/ServerException.java
@@ -1,0 +1,14 @@
+package ee.cyber.cdoc2.server.adapter.exception;
+
+public class ServerException extends RuntimeException {
+    private final String code;
+
+    public ServerException(String code, String message) {
+        super(message);
+        this.code = code;
+    }
+
+    public String getCode() {
+        return code;
+    }
+}